+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ### Unreleased
 
 - Added support for legacy basket format. ([#3](https://github.com/craftcms/commerce-sagepay/issues/3))
+- Added support for LOW profile to improve payment form display in iframes. ([#1](https://github.com/craftcms/commerce-sagepay/issues/1))
 
 ### 1.0.0